Module Ppx_protocol_driver.Default_parameters

Set of default Parameters

val field_name : string -> string

Map field names of records Mapping is done after applying [@key] attribute.

Default is identity

val variant_name : string -> string

Map variant and constructor names. Mapping is done after applying [@name] attribute.

Default is identity

val constructors_without_arguments_as_string : bool

Map constructors with no arguments to a string. If true, constructors without arguments are mapped to a string, instead of than a list containing only the constructor / variant name.

Default is true

val omit_default_values : bool

Omit default values from output. If true, default values are not serialized. Note that this uses polymorphic compare to determine if a field value is the same as the default value.

Default is true

val eager : bool

Lazy evaluate lazy fields. If true, lazy fields are parsed eagerly. If false, lazy fields are parsed first when forced, which means they will hold the serialized structure until forced, and forcing might raise a parse error.

Default is true

val strict : bool

Fail if unknown fields are encountered when deserialising records.

Default is false