Module Protocol_conv.Runtime

type nonrec ('v, 'e) result = ('v, 'e) Stdlib.result

module Record_in : sig ... end

module Record_out : sig ... end

module Tuple_in : sig ... end

module Tuple_out : sig ... end

module Variant_in : sig ... end

module type Driver = sig ... end

Signature for a driver. Serialization function are on the form of_XXX and deserialization function are on the form to_XXX.

module Helper : sig ... end

Module contains helper function for serializing and deserializing tuples, records and variants. Deserialization functions may raise Helper.Protocol exception. It is recommended that the calling functions convert this exception into a Driver.Protocol_exception